Allow importing library as ESM
