+++ .eslintrc.js
@@ -9,6 +9,7 @@ module.exports = {
   },
   parserOptions: {
     sourceType: 'module',
+    ecmaVersion: 2022,
   },
   rules: {
     'no-console': 'warn',
