+++ lib/.eslintrc.cjs
