+++ lib/handlebars/base.js
@@ -1,9 +1,11 @@
-import { Exception } from '@handlebars/parser';
-import { createFrame, extend, toString } from './utils';
-import { registerDefaultHelpers } from './helpers';
-import { registerDefaultDecorators } from './decorators';
-import logger from './logger';
-import { resetLoggedProperties } from './internal/proto-access';
+import * as parser from '@handlebars/parser';
+import { createFrame, extend, toString } from './utils.js';
+import { registerDefaultHelpers } from './helpers.js';
+import { registerDefaultDecorators } from './decorators.js';
+import logger from './logger.js';
+import { resetLoggedProperties } from './internal/proto-access.js';
+
+const Exception = parser.Exception ?? parser.default.Exception;
 
 export const VERSION = '4.7.7';
 export const COMPILER_REVISION = 8;
