+++ lib/handlebars/compiler/code-gen.js
@@ -1,50 +1,6 @@
-/* global define */
-import { isArray } from '../utils';
+import { SourceNode } from '#source-map';
 
-let SourceNode;
-
-try {
-  /* istanbul ignore next */
-  if (typeof define !== 'function' || !define.amd) {
-    // We don't support this in AMD environments. For these environments, we assume that
-    // they are running on the browser and thus have no need for the source-map library.
-    let SourceMap = require('source-map'); // eslint-disable-line no-undef
-    SourceNode = SourceMap.SourceNode;
-  }
-} catch (err) {
-  /* NOP */
-}
-
-/* istanbul ignore if: tested but not covered in istanbul due to dist build  */
-if (!SourceNode) {
-  SourceNode = function (line, column, srcFile, chunks) {
-    this.src = '';
-    if (chunks) {
-      this.add(chunks);
-    }
-  };
-  /* istanbul ignore next */
-  SourceNode.prototype = {
-    add: function (chunks) {
-      if (isArray(chunks)) {
-        chunks = chunks.join('');
-      }
-      this.src += chunks;
-    },
-    prepend: function (chunks) {
-      if (isArray(chunks)) {
-        chunks = chunks.join('');
-      }
-      this.src = chunks + this.src;
-    },
-    toStringWithSourceMap: function () {
-      return { code: this.toString() };
-    },
-    toString: function () {
-      return this.src;
-    },
-  };
-}
+import { isArray } from '../utils.js';
 
 function castChunk(chunk, codeGen, loc) {
   if (isArray(chunk)) {
