+++ lib/handlebars/compiler/compiler.js
@@ -1,8 +1,10 @@
 /* eslint-disable new-cap */
 
-import { Exception } from '@handlebars/parser';
-import { isArray, indexOf, extend } from '../utils';
-import AST from './ast';
+import * as parser from '@handlebars/parser';
+import { isArray, indexOf, extend } from '../utils.js';
+import AST from './ast.js';
+
+const Exception = parser.Exception ?? parser.default.Exception;
 
 const slice = [].slice;
