+++ lib/handlebars/compiler/javascript-compiler.js
@@ -1,7 +1,9 @@
-import { Exception } from '@handlebars/parser';
-import { COMPILER_REVISION, REVISION_CHANGES } from '../base';
-import { isArray } from '../utils';
-import CodeGen from './code-gen';
+import * as parser from '@handlebars/parser';
+import { COMPILER_REVISION, REVISION_CHANGES } from '../base.js';
+import { isArray } from '../utils.js';
+import CodeGen from './code-gen.js';
+
+const Exception = parser.Exception ?? parser.default.Exception;
 
 function Literal(value) {
   this.value = value;
