+++ lib/handlebars/decorators.js
@@ -1,4 +1,4 @@
-import registerInline from './decorators/inline';
+import registerInline from './decorators/inline.js';
 
 export function registerDefaultDecorators(instance) {
   registerInline(instance);
