+++ lib/handlebars/decorators/inline.js
@@ -1,4 +1,4 @@
-import { extend } from '../utils';
+import { extend } from '../utils.js';
 
 export default function (instance) {
   instance.registerDecorator(
