+++ lib/handlebars/helpers.js
@@ -1,10 +1,10 @@
-import registerBlockHelperMissing from './helpers/block-helper-missing';
-import registerEach from './helpers/each';
-import registerHelperMissing from './helpers/helper-missing';
-import registerIf from './helpers/if';
-import registerLog from './helpers/log';
-import registerLookup from './helpers/lookup';
-import registerWith from './helpers/with';
+import registerBlockHelperMissing from './helpers/block-helper-missing.js';
+import registerEach from './helpers/each.js';
+import registerHelperMissing from './helpers/helper-missing.js';
+import registerIf from './helpers/if.js';
+import registerLog from './helpers/log.js';
+import registerLookup from './helpers/lookup.js';
+import registerWith from './helpers/with.js';
 
 export function registerDefaultHelpers(instance) {
   registerBlockHelperMissing(instance);
