+++ lib/handlebars/helpers/block-helper-missing.js
@@ -1,4 +1,4 @@
-import { isArray } from '../utils';
+import { isArray } from '../utils.js';
 
 export default function (instance) {
   instance.registerHelper('blockHelperMissing', function (context, options) {
