+++ lib/handlebars/helpers/each.js
@@ -1,5 +1,7 @@
-import { Exception } from '@handlebars/parser';
-import { createFrame, isArray, isFunction, isMap, isSet } from '../utils';
+import * as parser from '@handlebars/parser';
+import { createFrame, isArray, isFunction, isMap, isSet } from '../utils.js';
+
+const Exception = parser.Exception ?? parser.default.Exception;
 
 export default function (instance) {
   instance.registerHelper('each', function (context, options) {
