+++ lib/handlebars/helpers/helper-missing.js
@@ -1,4 +1,6 @@
-import { Exception } from '@handlebars/parser';
+import * as parser from '@handlebars/parser';
+
+const Exception = parser.Exception ?? parser.default.Exception;
 
 export default function (instance) {
   instance.registerHelper('helperMissing', function (/* [args, ]options */) {
