+++ lib/handlebars/helpers/if.js
@@ -1,5 +1,7 @@
-import { Exception } from '@handlebars/parser';
-import { isEmpty, isFunction } from '../utils';
+import * as parser from '@handlebars/parser';
+import { isEmpty, isFunction } from '../utils.js';
+
+const Exception = parser.Exception ?? parser.default.Exception;
 
 export default function (instance) {
   instance.registerHelper('if', function (conditional, options) {
