+++ lib/handlebars/index.js
@@ -1,18 +1,18 @@
-import {
-  parser as Parser,
-  parse,
-  parseWithoutProcessing,
-  Visitor,
-} from '@handlebars/parser';
+import * as parser from '@handlebars/parser';
 
-import runtime from './handlebars.runtime';
+const { parse, parseWithoutProcessing } = parser;
+
+const Parser = parser.parser ?? parser.default.parser;
+const Visitor = parser.Visitor ?? parser.default.Visitor;
+
+import runtime from './index.runtime.js';
 
 // Compiler imports
-import AST from './handlebars/compiler/ast';
-import { Compiler, compile, precompile } from './handlebars/compiler/compiler';
-import JavaScriptCompiler from './handlebars/compiler/javascript-compiler';
+import AST from './compiler/ast.js';
+import { Compiler, compile, precompile } from './compiler/compiler.js';
+import JavaScriptCompiler from './compiler/javascript-compiler.js';
 
-import noConflict from './handlebars/no-conflict';
+import noConflict from './no-conflict.js';
 
 let _create = runtime.create;
 function create() {
