+++ lib/handlebars/index.runtime.js
@@ -1,13 +1,17 @@
-import { Exception } from '@handlebars/parser';
-import * as base from './handlebars/base';
+import * as parser from '@handlebars/parser';
+import * as base from './base.js';
 
 // Each of these augment the Handlebars object. No need to setup here.
 // (This is done to easily share code between commonjs and browse envs)
-import SafeString from './handlebars/safe-string';
-import * as Utils from './handlebars/utils';
-import * as runtime from './handlebars/runtime';
+import SafeString from './safe-string.js';
+import * as Utils from './utils.js';
+import * as runtime from './runtime.js';
 
-import noConflict from './handlebars/no-conflict';
+import noConflict from './no-conflict.js';
+
+const Exception = parser.Exception ?? parser.default.Exception;
+
+const vm = Utils.extend({}, runtime);
 
 // For compatibility and usage outside of module systems, make the Handlebars object a namespace
 function create() {
@@ -19,7 +23,7 @@ function create() {
   hb.Utils = Utils;
   hb.escapeExpression = Utils.escapeExpression;
 
-  hb.VM = runtime;
+  hb.VM = vm;
   hb.template = function (spec) {
     return runtime.template(spec, hb);
   };
