+++ lib/handlebars/internal/create-new-lookup-object.js
@@ -1,4 +1,4 @@
-import { extend } from '../utils';
+import { extend } from '../utils.js';
 
 /**
  * Create a new object with "null"-prototype to avoid truthy results on prototype properties.
