+++ lib/handlebars/internal/proto-access.js
@@ -1,5 +1,5 @@
-import { createNewLookupObject } from './create-new-lookup-object';
-import logger from '../logger';
+import { createNewLookupObject } from './create-new-lookup-object.js';
+import logger from '../logger.js';
 
 const loggedProperties = Object.create(null);
