+++ lib/handlebars/logger.js
@@ -1,4 +1,4 @@
-import { indexOf } from './utils';
+import { indexOf } from './utils.js';
 
 let logger = {
   methodMap: ['debug', 'info', 'warn', 'error'],
