+++ lib/handlebars/runtime.js
@@ -1,17 +1,19 @@
-import { Exception } from '@handlebars/parser';
-import * as Utils from './utils';
+import * as parser from '@handlebars/parser';
+import * as Utils from './utils.js';
 import {
   COMPILER_REVISION,
   createFrame,
   LAST_COMPATIBLE_COMPILER_REVISION,
   REVISION_CHANGES,
-} from './base';
-import { moveHelperToHooks } from './helpers';
-import { wrapHelper } from './internal/wrapHelper';
+} from './base.js';
+import { moveHelperToHooks } from './helpers.js';
+import { wrapHelper } from './internal/wrapHelper.js';
 import {
   createProtoAccessControl,
   resultIsAllowed,
-} from './internal/proto-access';
+} from './internal/proto-access.js';
+
+const Exception = parser.Exception ?? parser.default.Exception;
 
 export function checkRevision(compilerInfo) {
   const compilerRevision = (compilerInfo && compilerInfo[0]) || 1,
