+++ lib/index.cjs
@@ -1,0 +1,2 @@
+/* eslint-env node */
+module.exports = require('../dist/cjs/index.js').default;
