+++ lib/index.js
@@ -1,26 +1,7 @@
-// USAGE:
-// var handlebars = require('handlebars');
-/* eslint-env node */
-/* eslint-disable no-var */
+import handlebars from './handlebars/index.js';
+import { PrintVisitor, print } from '@handlebars/parser';
 
-// var local = handlebars.create();
+handlebars.PrintVisitor = PrintVisitor;
+handlebars.print = print;
 
-var handlebars = require('../dist/cjs/handlebars')['default'];
-
-var parser = require('@handlebars/parser');
-handlebars.PrintVisitor = parser.PrintVisitor;
-handlebars.print = parser.print;
-
-module.exports = handlebars;
-
-// Publish a Node.js require() handler for .handlebars and .hbs files
-function extension(module, filename) {
-  var fs = require('fs');
-  var templateString = fs.readFileSync(filename, 'utf8');
-  module.exports = handlebars.compile(templateString);
-}
-/* istanbul ignore else */
-if (typeof require !== 'undefined' && require.extensions) {
-  require.extensions['.handlebars'] = extension;
-  require.extensions['.hbs'] = extension;
-}
+export default handlebars;
