+++ lib/index.runtime.cjs
@@ -1,0 +1,2 @@
+/* eslint-env node */
+module.exports = require('../dist/cjs/handlebars/index.runtime.js').default;
