+++ lib/precompiler.js
@@ -1,8 +1,8 @@
 /* eslint-env node */
 /* eslint-disable no-console */
 import Async from 'neo-async';
-import fs from 'fs';
-import * as Handlebars from './handlebars';
+import * as fs from 'fs';
+import Handlebars from './handlebars/index.js';
 import { basename } from 'path';
 import { SourceMapConsumer, SourceNode } from 'source-map';
