--- runtime.d.ts
@@ -1,5 +1,0 @@
-import Handlebars = require('handlebars')
-
-declare module "handlebars/runtime" {
-   
-}
